Common FlowArrow
